test files moved to respective dirs
