+++ components/activity/description/test/d2l-activity-description-editor.test.js
@@ -1,5 +1,6 @@
+/* eslint-disable no-undef */
 import '../d2l-activity-description-editor.js';
-import { addToMock, mockLink } from '../../../../test/data/fetchMocks.js';
+import { addToMock, mockLink } from '../../../../test/data/fetchMock.js';
 import { assert, elementUpdated, expect, html } from '@open-wc/testing';
 import { createComponentAndWait, delayAndAwaitForElement, fireEventAndWait } from '../../../../test/test-util.js';
 import { learningPathExisting, learningPathMissingAction, learningPathNew, learningPathUpdated } from '../../../../test/data/learningPath.js';
